requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
ecdsa==0.18.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
python-dotenv==0.21.1
python-jose==3.3.0
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
